backend/library/requirements.txt
aiohttp==3.6.2
asgiref==3.5.0
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
backports.zoneinfo==0.2.1
certifi==2020.4.5.1
chardet==3.0.4
click==8.0.1
colorama==0.4.3
cycler==0.10.0
discord.py==1.3.3
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.6.0
get==2019.4.13
gunicorn==20.1.0
idna==2.9
isort==4.3.21
joblib==1.0.1
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
matplotlib==3.4.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.21.2
Pillow==8.3.2
pipenv==2022.1.8
platformdirs==2.5.1
ply==3.11
post==2019.4.13
psutil==5.8.0
psycopg2==2.9.3
public==2019.4.13
pycryptodome==3.10.1
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
query-string==2019.4.13
regex==2021.8.28
request==2019.4.13
requests==2.23.0
six==1.15.0
source==1.2.0
sqlparse==0.4.2
toml==0.10.1
tqdm==4.62.2
tzdata==2021.5
urllib3==1.25.9
virtualenv==20.13.1
virtualenv-clone==0.5.7
websockets==8.1
whitenoise==6.0.0
wrapt==1.12.1
yarl==1.4.2
